# Change all internal metrics to be automatically generated