--- a/internal/data_generator/internal/metrics_structs.go
+++ b/internal/data_generator/internal/metrics_structs.go
@@ -17,19 +17,147 @@ package internal
 var metricsFile = &File{
 	Name: "metrics",
 	imports: []string{
+		`otlpcommon "github.com/open-telemetry/opentelemetry-proto/gen/go/common/v1"`,
 		`otlpmetrics "github.com/open-telemetry/opentelemetry-proto/gen/go/metrics/v1"`,
+		`otlpresource "github.com/open-telemetry/opentelemetry-proto/gen/go/resource/v1"`,
 	},
 	structs: []baseStruct{
+		resourceMetricsSlice,
+		resourceMetrics,
+		instrumentationLibraryMetricsSlice,
+		instrumentationLibraryMetrics,
+		metricSlice,
+		metric,
+		metricDescriptor,
+		int64DataPointSlice,
 		int64DataPoint,
+		doubleDataPointSlice,
 		doubleDataPoint,
+		histogramDataPointSlice,
 		histogramDataPoint,
 		histogramBucketSlice,
 		histogramBucket,
 		histogramBucketExemplar,
+		summaryDataPointSlice,
 		summaryDataPoint,
 		summaryValueAtPercentileSlice,
 		summaryValueAtPercentile,
 	},
+}
+
+var resourceMetricsSlice = &sliceStruct{
+	structName: "ResourceMetricsSlice",
+	element:    resourceMetrics,
+}
+
+var resourceMetrics = &messageStruct{
+	structName:     "ResourceMetrics",
+	description:    "// InstrumentationLibraryMetrics is a collection of metrics from a LibraryInstrumentation.",
+	originFullName: "otlpmetrics.ResourceMetrics",
+	fields: []baseField{
+		resourceField,
+		&sliceField{
+			fieldMame:       "InstrumentationLibraryMetrics",
+			originFieldName: "InstrumentationLibraryMetrics",
+			returnSlice:     instrumentationLibraryMetricsSlice,
+		},
+	},
+}
+
+var instrumentationLibraryMetricsSlice = &sliceStruct{
+	structName: "InstrumentationLibraryMetricsSlice",
+	element:    instrumentationLibraryMetrics,
+}
+
+var instrumentationLibraryMetrics = &messageStruct{
+	structName:     "InstrumentationLibraryMetrics",
+	description:    "// InstrumentationLibraryMetrics is a collection of metrics from a LibraryInstrumentation.",
+	originFullName: "otlpmetrics.InstrumentationLibraryMetrics",
+	fields: []baseField{
+		instrumentationLibraryField,
+		&sliceField{
+			fieldMame:       "Metrics",
+			originFieldName: "Metrics",
+			returnSlice:     metricSlice,
+		},
+	},
+}
+
+var metricSlice = &sliceStruct{
+	structName: "MetricSlice",
+	element:    metric,
+}
+
+var metric = &messageStruct{
+	structName:     "Metric",
+	description:    "// MetricDescriptor is the descriptor of a metric.",
+	originFullName: "otlpmetrics.Metric",
+	fields: []baseField{
+		&messageField{
+			fieldMame:       "MetricDescriptor",
+			originFieldName: "MetricDescriptor",
+			returnMessage:   metricDescriptor,
+		},
+		&sliceField{
+			fieldMame:       "Int64DataPoints",
+			originFieldName: "Int64DataPoints",
+			returnSlice:     int64DataPointSlice,
+		},
+		&sliceField{
+			fieldMame:       "DoubleDataPoints",
+			originFieldName: "DoubleDataPoints",
+			returnSlice:     doubleDataPointSlice,
+		},
+		&sliceField{
+			fieldMame:       "HistogramDataPoints",
+			originFieldName: "HistogramDataPoints",
+			returnSlice:     histogramDataPointSlice,
+		},
+		&sliceField{
+			fieldMame:       "SummaryDataPoints",
+			originFieldName: "SummaryDataPoints",
+			returnSlice:     summaryDataPointSlice,
+		},
+	},
+}
+
+var metricDescriptor = &messageStruct{
+	structName:     "MetricDescriptor",
+	description:    "// MetricDescriptor is the descriptor of a metric.",
+	originFullName: "otlpmetrics.MetricDescriptor",
+	fields: []baseField{
+		&primitiveField{
+			fieldMame:       "Name",
+			originFieldName: "Name",
+			returnType:      "string",
+		},
+		&primitiveField{
+			fieldMame:       "Description",
+			originFieldName: "Description",
+			returnType:      "string",
+		},
+		&primitiveField{
+			fieldMame:       "Unit",
+			originFieldName: "Unit",
+			returnType:      "string",
+		},
+		&primitiveTypedField{
+			fieldMame:       "Type",
+			originFieldName: "Type",
+			returnType:      "MetricType",
+			rawType:         "otlpmetrics.MetricDescriptor_Type",
+		},
+		&sliceField{
+			fieldMame:       "LabelsMap",
+			originFieldName: "Labels",
+			returnSlice:     stringMap,
+		},
+	},
+}
+
+var int64DataPointSlice = &sliceStruct{
+	structName: "Int64DataPointSlice",
+	element:    int64DataPoint,
 }
 
 var int64DataPoint = &messageStruct{
@@ -52,6 +180,11 @@ var int64DataPoint = &messageStruct{
 	},
 }
 
+var doubleDataPointSlice = &sliceStruct{
+	structName: "DoubleDataPointSlice",
+	element:    doubleDataPoint,
+}
+
 var doubleDataPoint = &messageStruct{
 	structName:     "DoubleDataPoint",
 	description:    "// DoubleDataPoint is a single data point in a timeseries that describes the time-varying value of a double metric.",
@@ -70,6 +203,11 @@ var doubleDataPoint = &messageStruct{
 			returnType:      "float64",
 		},
 	},
+}
+
+var histogramDataPointSlice = &sliceStruct{
+	structName: "HistogramDataPointSlice",
+	element:    histogramDataPoint,
 }
 
 var histogramDataPoint = &messageStruct{
@@ -136,6 +274,11 @@ var histogramBucketExemplar = &messageStruct{
 			returnSlice:     stringMap,
 		},
 	},
+}
+
+var summaryDataPointSlice = &sliceStruct{
+	structName: "SummaryDataPointSlice",
+	element:    summaryDataPoint,
 }
 
 var summaryDataPoint = &messageStruct{