--- a/internal/data_generator/internal/resource_structs.go
+++ b/internal/data_generator/internal/resource_structs.go
@@ -36,3 +36,9 @@ var resource = &messageStruct{
 		},
 	},
 }
+
+var resourceField = &messageField{
+	fieldMame:       "Resource",
+	originFieldName: "Resource",
+	returnMessage:   resource,
+}