Before:
```
go test -benchmem -bench=. -run=notests ./internal/data/...                                                           
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector/internal/data
BenchmarkOtlpToFromInternal_PassThrough-16                              1000000000               0.262 ns/op           0 B/op          0 allocs/op
BenchmarkOtlpToFromInternal_Int64Points_MutateOneLabel-16                4814449               249 ns/op             240 B/op          7 allocs/op
BenchmarkOtlpToFromInternal_DoublePoints_MutateOneLabel-16               4815667               253 ns/op             240 B/op          7 allocs/op
BenchmarkOtlpToFromInternal_HistogramPoints_MutateOneLabel-16            4795528               251 ns/op             240 B/op          7 allocs/op
BenchmarkOtlpToFromInternal_SummaryPoints_MutateOneLabel-16              4808486               249 ns/op             240 B/op          7 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector/internal/data 6.415s
```

After:
```
go test -benchmem -bench=. -run=notests ./internal/data/...
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector/internal/data
BenchmarkOtlpToFromInternal_PassThrough-16                              1000000000               0.238 ns/op           0 B/op          0 allocs/op
BenchmarkOtlpToFromInternal_Int64Points_MutateOneLabel-16               141672010                8.15 ns/op            0 B/op          0 allocs/op
BenchmarkOtlpToFromInternal_DoublePoints_MutateOneLabel-16              146423392                8.17 ns/op            0 B/op          0 allocs/op
BenchmarkOtlpToFromInternal_HistogramPoints_MutateOneLabel-16           142867308                8.19 ns/op            0 B/op          0 allocs/op
BenchmarkOtlpToFromInternal_SummaryPoints_MutateOneLabel-16             145813492                8.18 ns/op            0 B/op          0 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector/internal/data 9.692s
```