@bogdandrutu 2020-03-22
The code coverage will get fixed after #679 is merged and code gets regenerated